fix(context): specified viewer should be prioritized the first one
