+++ app_test.go
@@ -610,148 +610,6 @@ func TestHtmlViewEngine(t *testing.T) {
 
 }
 
-func TestMixedViewers(t *testing.T) {
-	fsys := fstest.MapFS{
-		"views/user.html":  {Data: []byte(`user`)},
-		"pages/index.html": {Data: []byte(`index`)},
-		"pages/list.html":  {Data: []byte(`list`)},
-	}
-
-	mux := http.NewServeMux()
-	srv := httptest.NewServer(mux)
-	defer srv.Close()
-
-	app := New(WithMux(mux), WithFsys(fsys))
-
-	app.Get("/index", func(c *Context) error {
-		return c.View(nil, "index")
-	})
-
-	app.Get("/{$}", func(c *Context) error {
-		if c.Request().URL.Path == "/" {
-			return c.View(nil)
-		}
-
-		c.WriteStatus(http.StatusNotFound)
-		return ErrCancelled
-	})
-
-	app.Get("/list", func(c *Context) error {
-		return c.View(map[string]any{
-			"name": "list",
-			"num":  2,
-		})
-	})
-
-	app.Start()
-	defer app.Close()
-
-	req, err := http.NewRequest("GET", srv.URL+"/", nil)
-	req.Header.Set("Accept", "text/html, */*")
-	require.NoError(t, err)
-	resp, err := client.Do(req)
-	require.NoError(t, err)
-
-	buf, err := io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	require.Equal(t, fsys["pages/index.html"].Data, buf)
-
-	req, err = http.NewRequest("GET", srv.URL+"/", nil)
-	req.Header.Set("Accept", "text/html, */*")
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-
-	buf, err = io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	require.Equal(t, fsys["pages/index.html"].Data, buf)
-
-	req, err = http.NewRequest("GET", srv.URL+"/404", nil)
-	req.Header.Set("Accept", "text/html, */*")
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-	require.Equal(t, http.StatusNotFound, resp.StatusCode)
-	resp.Body.Close()
-
-	req, err = http.NewRequest("GET", srv.URL+"/list", nil)
-	req.Header.Set("Accept", "text/html, */*")
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-
-	buf, err = io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	require.Equal(t, fsys["pages/list.html"].Data, buf)
-
-	req, err = http.NewRequest("GET", srv.URL+"/list", nil)
-	req.Header.Set("Accept", "application/json")
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-
-	buf, err = io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	data := &struct {
-		Name string `json:"name"`
-		Num  int    `json:"num"`
-	}{}
-
-	err = json.Unmarshal(buf, data)
-
-	require.NoError(t, err)
-
-	require.Equal(t, data.Name, "list")
-	require.Equal(t, data.Num, 2)
-
-	// accept doesn't matched , first viewer(text/html) should be used
-	req, err = http.NewRequest("GET", srv.URL+"/", nil)
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-
-	buf, err = io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	require.Equal(t, fsys["pages/index.html"].Data, buf)
-
-	// getViewer should match its viewer
-	req, err = http.NewRequest("GET", srv.URL+"/index", nil)
-	req.Header.Set("Accept", "text/html")
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-
-	buf, err = io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	require.Equal(t, fsys["pages/index.html"].Data, buf)
-
-	// wildcard accept should be matched on first viewer
-	req, err = http.NewRequest("GET", srv.URL+"/", nil)
-	req.Header.Set("Accept", "*/*")
-	require.NoError(t, err)
-	resp, err = client.Do(req)
-	require.NoError(t, err)
-
-	buf, err = io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	resp.Body.Close()
-
-	require.Equal(t, fsys["pages/index.html"].Data, buf)
-
-}
-
 func TestDataBindOnHtml(t *testing.T) {
 	fsys := fstest.MapFS{
 		"pages/users.html": {Data: []byte(`<html><body>
